Control_panel
this repository is to control the robot arm movement by choosing the side to move through the simple interface which contains 5 buttons (Forward, left, backward, right and stop) where every button send a value to DB and print it in spreated page. I worked with Fantstic Rahaf from Fantastic Group.